changed some float and double usage to reduce picky compile warnings
git-svn-id: http://vtp.googlecode.com/svn/trunk@5146 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/Enviro/EarthView.cpp
@@ -552,7 +552,7 @@ void Enviro::DoCursorOnEarth()
 vtTerrain *Enviro::FindTerrainOnEarth(const DPoint2 &p)
 {
 	vtTerrain *t, *smallest = NULL;
-	float diag, smallest_diag = 1E7;
+	double diag, smallest_diag = 1E7;
 
 	for (unsigned int i = 0; i < NumTerrains(); i++)
 	{
